services/analyzer/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
anthropic==0.42.0
openai==1.58.1
python-dotenv==1.0.1
